requirements.txt
flask==3.1.0
flask-cors==4.0.0
pymongo==4.6.1
pyjwt==2.8.0
werkzeug>=3.1
requests==2.31.0 